# 18 format date with month and day options