--- a/cldr.go
+++ b/cldr.go
@@ -1975,16 +1975,70 @@ func fmtMonthDayGregorian(locale language.Tag, digits digits, opts Options) func
 	switch lang.String() {
 	default:
 		return func(m time.Month, d int) string { return fmtMonth(m, Month2Digit) + "-" + fmtDay(d, Day2Digit) }
-	case "wae":
+	case "af", "as", "ia", "ky", "mi", "rm", "tg", "wo":
+		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "-" + fmtMonth(m, Month2Digit) }
+	case "lv":
+		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "." + fmtMonth(m, Month2Digit) + "." }
+	case "az", "cv", "fo", "hy", "kk", "ku", "os", "tk", "tt", "uk":
+		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "." + fmtMonth(m, Month2Digit) }
+	case "br", "fr", "ga", "it", "jv", "kkj", "sc", "syr", "uz", "vec":
+		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "/" + fmtMonth(m, Month2Digit) }
+	case "nb", "nn", "no":
+		return func(m time.Month, d int) string { return fmtDay(d, DayNumeric) + "." + fmtMonth(m, MonthNumeric) + "." }
+	case "sq":
+		return func(m time.Month, d int) string { return fmtDay(d, DayNumeric) + "." + fmtMonth(m, MonthNumeric) }
+	case "es", "ti":
+		return func(m time.Month, d int) string { return fmtDay(d, DayNumeric) + "/" + fmtMonth(m, MonthNumeric) }
+	case "pcm":
 		return func(m time.Month, d int) string {
-			fmtMonth = fmtMonthName(locale.String(), "stand-alone", "abbreviated")
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + " /" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
+		}
+	case "fy", "kok", "nl", "ug":
+		return func(m time.Month, d int) string {
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "-" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
+		}
+	case "cs", "sk", "sl":
+		return func(m time.Month, d int) string {
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + ". " + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "."
+		}
+	case "de", "dsb", "fi", "gsw", "hsb", "is", "lb", "smn":
+		return func(m time.Month, d int) string {
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "." + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "."
+		}
+	case "be", "da", "et", "he", "jgo", "ka":
+		return func(m time.Month, d int) string {
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "." + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
+		}
+	case "agq", "ast", "bas", "bm", "ca", "cy", "dje", "doi", "dua", "dyo", "el", "ewo", "ff", "fur", "gd", "gl", "haw", "hi", "id", "ig", "kab", "kgp", "khq", "km", "ksf", "kxv", "ln", "lo", "lu", "mai", "mfe", "mg", "mgh", "ml", "mni", "mua", "my", "nmg", "nus", "pa", "rn", "sa", "seh", "ses", "sg", "shi", "su", "sw", "to", "tr", "twq", "ur", "xnr", "yav", "yo", "yrl", "zgh":
+		return func(m time.Month, d int) string {
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "/" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
+		}
+	case "ar":
+		return func(m time.Month, d int) string {
+			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "‏/" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
+		}
+	case "iu":
+		return func(m time.Month, d int) string { return fmtMonth(m, Month2Digit) + "/" + fmtDay(d, Day2Digit) }
+	case "dz", "si":
+		return func(m time.Month, d int) string {
+			return fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "-" + fmtDay(d, cmp.Or(opts.Day, DayNumeric))
+		}
+	case "hu", "ko":
+		return func(m time.Month, d int) string {
+			return fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + ". " + fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "."
+		}
+	case "ak", "am", "asa", "bem", "bez", "blo", "brx", "ceb", "cgg", "chr", "dav", "ebu", "ee", "en", "eu", "fil", "guz", "ha", "ja", "jmc", "kam", "kde", "ki", "kln", "ks", "ksb", "lag", "lg", "luo", "luy", "mas", "mer", "naq", "nd", "nyn", "rof", "rwk", "saq", "sbp", "so", "teo", "tzm", "vai", "vun", "xh", "xog", "yue", "zh":
+		return func(m time.Month, d int) string {
+			return fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "/" + fmtDay(d, cmp.Or(opts.Day, DayNumeric))
+		}
+	case "wae":
+		fmtMonth = fmtMonthName(locale.String(), "stand-alone", "abbreviated")
+		return func(m time.Month, d int) string {
 			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + ". " + fmtMonth(m, opts.Month)
 		}
 	case "mn":
-		return func(m time.Month, d int) string {
-			fmtMonth = fmtMonthName(locale.String(), "stand-alone", "narrow")
-			return fmtMonth(m, opts.Month) + "/" + fmtDay(d, Day2Digit)
-		}
+		fmtMonth = fmtMonthName(locale.String(), "stand-alone", "narrow")
+		return func(m time.Month, d int) string { return fmtMonth(m, opts.Month) + "/" + fmtDay(d, Day2Digit) }
 	case "sv":
 		return func(m time.Month, d int) string {
 			if opts.Month == Month2Digit && opts.Day == DayNumeric {
@@ -2062,69 +2116,13 @@ func fmtMonthDayGregorian(locale language.Tag, digits digits, opts Options) func
 			}
 			return fmtDay(d, cmp.Or(opts.Day, Day2Digit)) + "-" + fmtMonth(m, cmp.Or(opts.Month, Month2Digit))
 		}
-	case "af", "as", "ia", "ky", "mi", "rm", "tg", "wo":
-		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "-" + fmtMonth(m, Month2Digit) }
-	case "az", "cv", "fo", "hy", "kk", "ku", "os", "tk", "tt", "uk":
-		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "." + fmtMonth(m, Month2Digit) }
-	case "lv":
-		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "." + fmtMonth(m, Month2Digit) + "." }
-	case "br", "fr", "ga", "it", "jv", "kkj", "sc", "syr", "uz", "vec":
-		return func(m time.Month, d int) string { return fmtDay(d, Day2Digit) + "/" + fmtMonth(m, Month2Digit) }
-	case "sq":
-		return func(m time.Month, d int) string { return fmtDay(d, DayNumeric) + "." + fmtMonth(m, MonthNumeric) }
-	case "nb", "nn", "no":
-		return func(m time.Month, d int) string { return fmtDay(d, DayNumeric) + "." + fmtMonth(m, MonthNumeric) + "." }
 	case "mk":
 		return func(m time.Month, d int) string {
 			return fmtDay(d, DayNumeric) + "." + fmtMonth(m, cmp.Or(opts.Month, Month2Digit))
 		}
-	case "es", "ti":
-		return func(m time.Month, d int) string { return fmtDay(d, DayNumeric) + "/" + fmtMonth(m, MonthNumeric) }
 	case "bg", "pl":
 		return func(m time.Month, d int) string {
 			return fmtDay(d, cmp.Or(opts.Day, Day2Digit)) + "." + fmtMonth(m, Month2Digit)
-		}
-	case "pcm":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + " /" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
-		}
-	case "fy", "kok", "nl", "ug":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "-" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
-		}
-	case "cs", "sk", "sl":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + ". " + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "."
-		}
-	case "be", "da", "et", "he", "jgo", "ka":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "." + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
-		}
-	case "de", "dsb", "fi", "gsw", "hsb", "is", "lb", "smn":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "." + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "."
-		}
-	case "agq", "ast", "bas", "bm", "ca", "cy", "dje", "doi", "dua", "dyo", "el", "ewo", "ff", "fur", "gd", "gl", "haw", "hi", "id", "ig", "kab", "kgp", "khq", "km", "ksf", "kxv", "ln", "lo", "lu", "mai", "mfe", "mg", "mgh", "ml", "mni", "mua", "my", "nmg", "nus", "pa", "rn", "sa", "seh", "ses", "sg", "shi", "su", "sw", "to", "tr", "twq", "ur", "xnr", "yav", "yo", "yrl", "zgh":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "/" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
-		}
-	case "ar":
-		return func(m time.Month, d int) string {
-			return fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "‏/" + fmtMonth(m, cmp.Or(opts.Month, MonthNumeric))
-		}
-	case "iu":
-		return func(m time.Month, d int) string { return fmtMonth(m, Month2Digit) + "/" + fmtDay(d, Day2Digit) }
-	case "dz", "si":
-		return func(m time.Month, d int) string {
-			return fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "-" + fmtDay(d, cmp.Or(opts.Day, DayNumeric))
-		}
-	case "hu", "ko":
-		return func(m time.Month, d int) string {
-			return fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + ". " + fmtDay(d, cmp.Or(opts.Day, DayNumeric)) + "."
-		}
-	case "ak", "am", "asa", "bem", "bez", "blo", "brx", "ceb", "cgg", "chr", "dav", "ebu", "ee", "en", "eu", "fil", "guz", "ha", "ja", "jmc", "kam", "kde", "ki", "kln", "ks", "ksb", "lag", "lg", "luo", "luy", "mas", "mer", "naq", "nd", "nyn", "rof", "rwk", "saq", "sbp", "so", "teo", "tzm", "vai", "vun", "xh", "xog", "yue", "zh":
-		return func(m time.Month, d int) string {
-			return fmtMonth(m, cmp.Or(opts.Month, MonthNumeric)) + "/" + fmtDay(d, cmp.Or(opts.Day, DayNumeric))
 		}
 	}
 }