--- a/go.mod
+++ b/go.mod
@@ -3,6 +3,7 @@ module go.expect.digital/intl
 go 1.23
 
 require (
+	github.com/google/go-cmp v0.6.0
 	github.com/yaa110/go-persian-calendar v1.2.1
 	golang.org/x/text v0.20.0
 )