--- a/internal/gen/cldr.go.tmpl
+++ b/internal/gen/cldr.go.tmpl
@@ -12,7 +12,6 @@ import (
 
 type calendarMonths [12]string
 
-// size 144384
 var calendarMonthNames = [...]calendarMonths{
 {{- range $calendarMonths := .Months.List }}
   {{ $calendarMonths }},