… go.sum
… internal/gen/gen_test.go
… internal/gen/main.go
… internal/gen/patterns.go
… internal/gen/patterns_test.go
… intl.go
… intl_test.go
… testdata.js