```console
$ benchstat a.txt b.txt                         
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                         │    a.txt    │               b.txt                │
                         │   sec/op    │   sec/op     vs base               │
NewDateTime/fa-IR-12       66.24n ± 1%   69.33n ± 1%  +4.65% (p=0.000 n=10)
NewDateTime/lv-LV-12       104.0n ± 1%   104.8n ± 1%  +0.82% (p=0.030 n=10)
NewDateTime/dz-BT-12       105.5n ± 1%   108.3n ± 1%  +2.65% (p=0.000 n=10)
DateTime_Format/fa-IR-12   573.6n ± 1%   611.5n ± 2%  +6.61% (p=0.000 n=10)
DateTime_Format/lv-LV-12   140.3n ± 1%   150.2n ± 0%  +7.02% (p=0.000 n=10)
DateTime_Format/dz-BT-12   240.6n ± 0%   259.8n ± 0%  +7.94% (p=0.000 n=10)
geomean                    155.4n        163.0n       +4.92%

                         │   a.txt    │                b.txt                │
                         │    B/op    │    B/op     vs base                 │
NewDateTime/fa-IR-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12   72.00 ± 0%   72.00 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   32.00 ± 0%   32.00 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   64.00 ± 0%   64.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    16.25        16.25       +0.00%
¹ all samples are equal

                         │   a.txt    │                b.txt                │
                         │ allocs/op  │ allocs/op   vs base                 │
NewDateTime/fa-IR-12       1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12       1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12   9.000 ± 0%   9.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   8.000 ± 0%   8.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    2.570        2.570       +0.00%
¹ all samples are equal
```